+++ tests/test_prg_engine_runtime_surface_functions_tests.h
@@ -20,7 +20,13 @@ namespace copperfin::runtime_surface_tests
     void test_lupdate_reads_table_header_date_and_designator_variants();
     void test_lupdate_respects_set_date_format();
     void test_lupdate_designator_edge_cases();
-    void test_curval_verified_commit_overlay_does_not_survive_cursor_reopen();
+    void test_curval_verified_commit_overlay_survives_cursor_reopen();
+    void test_curval_verified_admission_patch_tracks_second_cursor_memo_and_rollback();
+    void test_curval_verified_rollback_preserves_case_distinct_posix_admissions();
+    void test_curval_verified_command_undo_restores_admission_bytes();
+    void test_curval_verified_failed_command_undo_keeps_admission_bytes();
+    void test_curval_verified_force_commit_merges_other_alias_admission();
+    void test_curval_verified_rollback_deduplicates_windows_case_aliases();
     void test_getfldstate_tracks_buffered_mutation_state();
     void test_setfldstate_assigns_buffered_mutation_state();
     void test_oldval_evaluates_buffered_original_record();